[PM-37768] Move WebAuthn Permissions Policy gate to content script
