+++ apps/browser/src/autofill/fido2/content/fido2-page-script.webauthn-supported.spec.ts
@@ -143,141 +143,6 @@ describe("Fido2 page script with native WebAuthn support", () => {
     });
   });
 
-  describe("Permissions Policy enforcement", () => {
-    beforeEach(() => {
-      messenger.request = jest.fn();
-    });
-
-    afterEach(() => {
-      delete (mockGlobalThisDocument as any).featurePolicy;
-      delete (mockGlobalThisDocument as any).permissionsPolicy;
-    });
-
-    it("throws NotAllowedError without messaging the background when publickey-credentials-create is denied", async () => {
-      (mockGlobalThisDocument as any).featurePolicy = {
-        allowsFeature: jest.fn((feature: string) => feature !== "publickey-credentials-create"),
-      };
-
-      await expect(navigator.credentials.create(mockCredentialCreationOptions)).rejects.toEqual(
-        expect.objectContaining({
-          name: "NotAllowedError",
-        }),
-      );
-      expect(messenger.request).not.toHaveBeenCalled();
-    });
-
-    it("throws NotAllowedError without messaging the background when publickey-credentials-get is denied", async () => {
-      (mockGlobalThisDocument as any).featurePolicy = {
-        allowsFeature: jest.fn((feature: string) => feature !== "publickey-credentials-get"),
-      };
-
-      await expect(navigator.credentials.get(mockCredentialRequestOptions)).rejects.toEqual(
-        expect.objectContaining({
-          name: "NotAllowedError",
-        }),
-      );
-      expect(messenger.request).not.toHaveBeenCalled();
-    });
-
-    it("proceeds when the document's Permissions Policy allows the WebAuthn feature", async () => {
-      (mockGlobalThisDocument as any).featurePolicy = {
-        allowsFeature: jest.fn(() => true),
-      };
-      messenger.request = jest.fn().mockResolvedValue({
-        type: MessageTypes.CredentialGetResponse,
-        result: mockCredentialAssertResult,
-      });
-
-      await navigator.credentials.get(mockCredentialRequestOptions);
-
-      expect(messenger.request).toHaveBeenCalled();
-    });
-
-    it("prefers document.permissionsPolicy over document.featurePolicy when both are present", async () => {
-      const permissionsAllowsFeature = jest.fn(() => false);
-      const featureAllowsFeature = jest.fn(() => true);
-      (mockGlobalThisDocument as any).permissionsPolicy = {
-        allowsFeature: permissionsAllowsFeature,
-      };
-      (mockGlobalThisDocument as any).featurePolicy = {
-        allowsFeature: featureAllowsFeature,
-      };
-
-      await expect(navigator.credentials.get(mockCredentialRequestOptions)).rejects.toEqual(
-        expect.objectContaining({ name: "NotAllowedError" }),
-      );
-      expect(permissionsAllowsFeature).toHaveBeenCalledWith("publickey-credentials-get");
-      expect(featureAllowsFeature).not.toHaveBeenCalled();
-      expect(messenger.request).not.toHaveBeenCalled();
-    });
-
-    it("uses document.permissionsPolicy when featurePolicy is absent", async () => {
-      (mockGlobalThisDocument as any).permissionsPolicy = {
-        allowsFeature: jest.fn((feature: string) => feature !== "publickey-credentials-create"),
-      };
-
-      await expect(navigator.credentials.create(mockCredentialCreationOptions)).rejects.toEqual(
-        expect.objectContaining({ name: "NotAllowedError" }),
-      );
-      expect(messenger.request).not.toHaveBeenCalled();
-    });
-  });
-
-  describe("Permissions Policy enforcement — defense-in-depth fallback", () => {
-    let topSpy: jest.SpyInstance | undefined;
-    let selfSpy: jest.SpyInstance | undefined;
-
-    beforeEach(() => {
-      messenger.request = jest.fn();
-    });
-
-    afterEach(() => {
-      topSpy?.mockRestore();
-      selfSpy?.mockRestore();
-      topSpy = undefined;
-      selfSpy = undefined;
-    });
-
-    it("denies when no Permissions Policy API is available and we're in a cross-origin iframe", async () => {
-      const fakeSelf = { location: { origin: "https://child.example" } } as Window;
-      const fakeTop = { location: { origin: "https://parent.example" } } as Window;
-      selfSpy = jest.spyOn(globalThis, "self", "get").mockReturnValue(fakeSelf as any);
-      topSpy = jest.spyOn(globalThis, "top", "get").mockReturnValue(fakeTop as any);
-
-      await expect(navigator.credentials.get(mockCredentialRequestOptions)).rejects.toEqual(
-        expect.objectContaining({ name: "NotAllowedError" }),
-      );
-      expect(messenger.request).not.toHaveBeenCalled();
-    });
-
-    it("denies when no Permissions Policy API is available and accessing top.location throws", async () => {
-      const fakeSelf = { location: { origin: "https://child.example" } } as Window;
-      const fakeTop = {
-        get location() {
-          throw new DOMException("blocked", "SecurityError");
-        },
-      } as unknown as Window;
-      selfSpy = jest.spyOn(globalThis, "self", "get").mockReturnValue(fakeSelf as any);
-      topSpy = jest.spyOn(globalThis, "top", "get").mockReturnValue(fakeTop as any);
-
-      await expect(navigator.credentials.create(mockCredentialCreationOptions)).rejects.toEqual(
-        expect.objectContaining({ name: "NotAllowedError" }),
-      );
-      expect(messenger.request).not.toHaveBeenCalled();
-    });
-
-    it("allows when no Permissions Policy API is available and we're top-level (self === top)", async () => {
-      messenger.request = jest.fn().mockResolvedValue({
-        type: MessageTypes.CredentialGetResponse,
-        result: mockCredentialAssertResult,
-      });
-
-      await navigator.credentials.get(mockCredentialRequestOptions);
-
-      expect(messenger.request).toHaveBeenCalled();
-    });
-  });
-
   describe("destroy", () => {
     it("should destroy the message listener when receiving a disconnect request", async () => {
       jest.spyOn(globalThis.top, "removeEventListener");
